feat(6projects, index): retrieved cards from contentful and displayed on index
Substituted the content management system of the cards so as to facilitate future edits. I still
need to implement support for links, spanish (I still need to decide if I'm going to maintain
support for spanish, given that Contentful only supports two locales and an alternative solution
would not be ideal) and the division by categories.